+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ### Changelog
 
-**0.86.x**
+**0.86.6**
 
 * Minor updates
+
+**0.86.5**
+
+* Project Editor - Add Fields
+* Project Editor - Loading feature is added (Get project list of current user) 
 
 **0.86.4**